feat: execute cache_table_create in create_frontend_cache
