+++ inc/frontend_cache.h
@@ -8,16 +8,13 @@ t_cache* create_frontend_cache(t_pagemap* pagemap);
 t_tiny_chunk** create_fast_cache(t_cache* cache);
 
 size_t round_up_to_next(size_t number);
-t_chunk** create_cache_table(t_cache* cache);
 
-// unsigned int my_hash_function(size_t data_size, uint32_t table_size);
-size_t round_up_to_next(size_t number);
-t_chunk** create_cache_table(t_cache* cache);
-// void cache_table_destroy(t_cache_table* ht);
-// void cache_table_print(t_cache_table* ht);
-// bool cache_table_insert(t_heap **heap, t_cache_table *ht, const char *key, size_t value);
-// void* cache_table_allocate(t_cache_table* ht, size_t key);
-// void *cache_table_deallocate(t_block *block);
-// void *cache_table_deallocate(t_cache_table *ht, const char *key);
+cache_table* cache_table_create(t_cache* cache);
+void cache_table_destroy(cache_table* table);
+void* cache_table_get(cache_table* table, const char* key);
+const char* cache_table_set(cache_table* table, const char* key, t_chunk* value);
+size_t cache_table_length(cache_table* table);
+cache_tablei cache_table_iterator(cache_table* table);
+bool cache_table_next(cache_tablei* it);
 
 #endif
