+++ inc/utils.h
@@ -12,4 +12,6 @@ void system_settings();
 void check_system_pointer();
 size_t check_system_size_t();
 
+size_t round_up_to_next(size_t number);
+
 #endif
